image upload and ui changes

+++ acdesigns/dev-server/api/design/design-controller.js
@@ -2,6 +2,8 @@ import Design from '../../models/design-model'
 import aws from '../../config/aws'
 import * as auth from '../../services/auth-service'
 import fs from 'fs'
+import sharp from 'sharp'
+sharp.cache({ files : 0 });
 
 export function getDesign(req, res) {
     Design.findOne({ _id: req.params.id }, (error, design) => {
@@ -62,39 +64,49 @@ export function updateDesign(req, res) {
                 fs.unlinkSync(req.file.path) //empty uploads folder
                 return res.status(422).send('File must be of type .jpeg or .png') //status: Unprocessable Entity
             }
+            //update the image version (to refresh the cache)
+            design.imageVersion++
 
-            //upload the image to s3
-            const s3 = new aws.S3();
-            var params = {
-                ACL: 'public-read',
-                Bucket: process.env.BUCKET_NAME,
-                Body: fs.createReadStream(req.file.path),
-                Key: `designImage/${design._id}`,
-                ContentType: req.file.mimetype
-            };
-            s3.upload(params, (err, data) => {
-                if (err) {
-                    fs.unlinkSync(req.file.path) //empty uploads folder
-                    return res.status(500).send('Error uploading file') //status: internal server error
-                }
-                if (data) {
-                    fs.unlinkSync(req.file.path) //empty uploads folder
-                    //save the design to the db
-                    design.image = data.Location
-                    Design.findOneAndUpdate({ _id: design._id }, { $set: design }, (error) => {
-                        if (error) {
-                            return res.status(500).send('Error creating design') //status: internal server error
+            //resize the image
+            sharp(req.file.path).resize(500, 281).jpeg({quality: 90}).toBuffer()
+                .then(buff => {
+                    //upload the image to s3
+                    const s3 = new aws.S3();
+                    var params = {
+                        ACL: 'public-read',
+                        Bucket: process.env.BUCKET_NAME,
+                        Body: buff,
+                        Key: `designImage/${design._id}`,
+                        ContentType: 'image/jpeg'
+                    };
+                    s3.upload(params, (err, data) => {
+                        if (err) {
+                            fs.unlinkSync(req.file.path) //empty uploads folder
+                            return res.status(500).send('Error uploading file') //status: internal server error
                         }
-                        return res.status(200).json({ id: design._id }) //status: success
+                        if (data) {
+                            fs.unlinkSync(req.file.path) //empty uploads folder
+                            //save the design to the db
+                            design.image = data.Location
+                            Design.findOneAndUpdate({ _id: design._id }, { $set: design }, (error) => {
+                                if (error) {
+                                    return res.status(500).send('Error updating design') //status: internal server error
+                                }
+                                return res.status(200).json({ id: design._id }) //status: success
+                            })
+                        }
                     })
-                }
-            })
+                })
+                .catch(() => { //if sharp fails
+                    if (req.file) fs.unlinkSync(req.file.path) //empty uploads folder
+                    return res.status(500).send('Error updating design') //status: internal server error
+                })
         }
         //update the design without uploading a new image
         else {
             Design.findOneAndUpdate({ _id: design._id }, { $set: design }, (error) => {
                 if (error) {
-                    return res.status(500).send('Error creating design') //status: internal server error
+                    return res.status(500).send('Error updating design') //status: internal server error
                 }
                 return res.status(200).json({ id: design._id }) //status: success
             })
@@ -102,7 +114,7 @@ export function updateDesign(req, res) {
     })
         .catch(() => { //if unable to get users id
             if (req.file) fs.unlinkSync(req.file.path) //empty uploads folder
-            return res.status(500).send('Error creating design') //status: internal server error
+            return res.status(500).send('Error updating design') //status: internal server error
         })
 }
 
@@ -123,6 +135,7 @@ export function createDesign(req, res) {
             user: userId,
             designCode: req.body.designCode,
             image: '',
+            imageVersion: 1,
             title: req.body.title,
             description: req.body.description,
             type: req.body.type,
@@ -131,32 +144,40 @@ export function createDesign(req, res) {
         })
         const designId = design._id
 
-        //upload the image to s3
-        const s3 = new aws.S3();
-        var params = {
-            ACL: 'public-read',
-            Bucket: process.env.BUCKET_NAME,
-            Body: fs.createReadStream(req.file.path),
-            Key: `designImage/${designId}`,
-            ContentType: req.file.mimetype
-        };
-        s3.upload(params, (err, data) => {
-            if (err) {
-                fs.unlinkSync(req.file.path) //empty uploads folder
-                return res.status(500).send('Error uploading file') //status: internal server error
-            }
-            if (data) {
-                fs.unlinkSync(req.file.path) //empty uploads folder
-                //save the design to the db
-                design.image = data.Location
-                design.save(error => {
-                    if (error) {
-                        return res.status(500).send('Error creating design') //status: internal server error
+        //resize the image
+        sharp(req.file.path).resize(500, 281).jpeg({quality: 90}).toBuffer()
+            .then(buff => {
+                //upload the image to s3
+                const s3 = new aws.S3();
+                var params = {
+                    ACL: 'public-read',
+                    Bucket: process.env.BUCKET_NAME,
+                    Body: buff,
+                    Key: `designImage/${designId}`,
+                    ContentType: 'image/jpeg'
+                };
+                s3.upload(params, (err, data) => {
+                    if (err) {
+                        fs.unlinkSync(req.file.path) //empty uploads folder
+                        return res.status(500).send('Error uploading file') //status: internal server error
                     }
-                    return res.status(201).json({ id: designId }) //status: success, created
+                    if (data) {
+                        fs.unlinkSync(req.file.path) //empty uploads folder
+                        //save the design to the db
+                        design.image = data.Location
+                        design.save(error => {
+                            if (error) {
+                                return res.status(500).send('Error creating design') //status: internal server error
+                            }
+                            return res.status(201).json({ id: designId }) //status: success, created
+                        })
+                    }
                 })
-            }
-        })
+            })
+            .catch(() => { //if sharp fails
+                if (req.file) fs.unlinkSync(req.file.path) //empty uploads folder
+                return res.status(500).send('Error creating design') //status: internal server error
+            })
     })
         .catch(() => { //if unable to get users id
             if (req.file) fs.unlinkSync(req.file.path) //empty uploads folder
